Move to /model/entity
