+++ app/src/main/java/com/alongo/manyweather/data/model/entity/Coordinate.kt
@@ -1,3 +1,3 @@
-package com.alongo.manyweather.data.model
+package com.alongo.manyweather.data.model.entity
 
 data class Coordinate(val lat: Double, val lon: Double)
