+++ app/src/main/java/com/alongo/manyweather/data/model/entity/weather/WeatherInternalParameter.kt
@@ -1,4 +1,4 @@
-package com.alongo.manyweather.data.model.weather
+package com.alongo.manyweather.data.model.entity.weather
 
 data class WeatherInternalParameter(
     val type: Int,
